+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 2021-10-22
 ### Added
 - added `Folio::FriendlyIdForTraco` concern
+- added `Folio::HasAncestrySlugForTraco` concern
 
 ## 2021-10-21
 ### Added